WIP: modem-manager: implement Quectel EM120 firmware upgrade support
The Quectel EM120 is a PCI module that can be upgraded using a
combination of MHI/BHI (Modem Host Interface / Boot Host Interface)
and Qualcomm Firehose protocol.

The MHI/BHI step is used to upload to the module the Firehose
programmer implementation, and once it's ready, the Firehose protocol
is used to erase and upload new system partition files.

The MHI/BHI operation relies on the existence of a /dev/mhi_BHI
control port exposed by the modem, which is always available
regardless of the port layout it exposes. i.e. the mhi_BHI port is
available when the device is in 'modem' mode and also when the device
is in 'embedded downloader' mode. This mhi_BHI port is currently a
Quectel specific implementation in their pcie_mhi driver.

The Firehose operation relies on a manifest file that contains the
actions to be performed via the Firehose protocol, very similar to the
Fastboot plugin.

**********************************************************************
  The MHI_BHI and FIREHOSE update methods in ModemManager are NOT
  YET AVAILABLE in the upstream git master branch.
**********************************************************************